[zzawang] step-6, 7 ajax 댓글 구현

+++ src/main/java/codesquad/springcafe/controller/ArticleController.java
@@ -2,13 +2,17 @@ package codesquad.springcafe.controller;
 
 import codesquad.springcafe.dto.ArticleUpdateDto;
 import codesquad.springcafe.dto.ArticleWriteDto;
-import codesquad.springcafe.exception.service.DuplicateUserIdException;
+import codesquad.springcafe.dto.CommentWriteDto;
+import codesquad.springcafe.exception.service.DuplicateArticleIdException;
 import codesquad.springcafe.model.Article;
+import codesquad.springcafe.model.Comment;
 import codesquad.springcafe.model.SessionUser;
 import codesquad.springcafe.model.UpdatedArticle;
 import codesquad.springcafe.service.ArticleService;
+import codesquad.springcafe.service.CommentService;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,10 +33,12 @@ public class ArticleController {
     private static final Logger logger = LoggerFactory.getLogger(ArticleController.class);
 
     private final ArticleService articleService;
+    private final CommentService commentService;
 
     @Autowired
-    public ArticleController(ArticleService articleService) {
+    public ArticleController(ArticleService articleService, CommentService commentService) {
         this.articleService = articleService;
+        this.commentService = commentService;
     }
 
     @GetMapping("/write")
@@ -54,23 +60,44 @@ public class ArticleController {
             articleService.addArticle(article);
             sessionUser.addArticleId(article.getId()); // 세션 유저에 작성한 게시글 Id 추가
             return "redirect:/";
-        } catch (DuplicateUserIdException e) {
+        } catch (DuplicateArticleIdException e) {
             bindingResult.reject("errorAid");
             return "article/form";
         }
     }
 
-    @GetMapping("/show/{articleId}")
-    public String showDetailPage(@PathVariable long articleId, Model model, HttpSession httpSession) {
-        articleService.increaseViewCount(articleId);
-        Article article = articleService.findArticleById(articleId);
-
+    @GetMapping("/{articleId}")
+    public String showDetailPage(@PathVariable long articleId, Model model, HttpSession httpSession,
+                                 @ModelAttribute("commentWriteDto") CommentWriteDto commentWriteDto) {
+        canIncreaseViewCount(httpSession, articleId);
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("sessionUser");
-        boolean isWriter = sessionUser.getArticleIds().contains(articleId);
+        String userId = sessionUser.getUserId();
+        Article article = articleService.findArticleById(articleId);
+        List<Comment> comments = commentService.findCommentsByAid(articleId);
+        boolean canDelete = comments.stream().allMatch(comment -> comment.getUserId().equals(userId));
 
         model.addAttribute("article", article);
-        model.addAttribute("isWriter", isWriter);
-        return "article/show";
+        model.addAttribute("isWriter", sessionUser.getArticleIds().contains(articleId));
+        model.addAttribute("canDelete", canDelete);
+        model.addAttribute("writer", sessionUser.getUserId());
+        model.addAttribute("commentList", comments);
+
+        Object commentBlankError = httpSession.getAttribute("commentBlankError");
+        if (commentBlankError != null && (boolean) commentBlankError) {
+            model.addAttribute("IsCommentBlank", true);
+        }
+        httpSession.removeAttribute("commentBlankError");
+        httpSession.removeAttribute("commentControlFlag");
+        return "article/show_comment";
+    }
+
+    private void canIncreaseViewCount(HttpSession httpSession, long articleId) {
+        Object commentBlankError = httpSession.getAttribute("commentBlankError");
+        Object commentAddFlag = httpSession.getAttribute("commentControlFlag");
+
+        if (commentBlankError == null && commentAddFlag == null) {
+            articleService.increaseViewCount(articleId);
+        }
     }
 
     @GetMapping("/update/{articleId}")
@@ -90,21 +117,36 @@ public class ArticleController {
             return "article/update";
         }
 
-        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("sessionUser");
-        UpdatedArticle updatedArticle = articleUpdateDto.createUpdatedArticle(sessionUser.getUserId());
+        UpdatedArticle updatedArticle = articleUpdateDto.createUpdatedArticle();
         articleService.updateArticle(articleId, updatedArticle);
-        return "redirect:/";
+        httpSession.setAttribute("commentControlFlag", true);
+        return "redirect:/article/{articleId}";
     }
 
-    @DeleteMapping("/delete/{articleId}")
-    public String processDelete(@PathVariable long articleId) {
-        articleService.deleteArticle(articleId);
-        return "redirect:/";
+    @DeleteMapping("/update/{articleId}")
+    public String processDelete(@PathVariable long articleId, HttpSession httpSession) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("sessionUser");
+        String userId = sessionUser.getUserId();
+        List<Comment> otherComments = commentService.findCommentsByAid(articleId)
+                .stream().filter(comment -> !comment.getUserId().equals(userId)).toList();
+
+        if (otherComments.isEmpty()) { // 다른 사용자의 댓글이 없는 경우만 게시글 삭제가 가능하다.
+            commentService.deleteArticlesComment(articleId);
+            articleService.deleteArticle(articleId);
+            return "redirect:/";
+        }
+        return "redirect:/article/invalid-delete";
     }
 
     @GetMapping("/invalid-modify")
     public String showInvalidModifyPage(Model model) {
-        model.addAttribute("errorMsg", "다른 사람의 게시글은 수정할 수 없습니다.");
+        model.addAttribute("errorMsg", "다른 사용자의 게시글과 댓글은 수정할 수 없습니다.");
+        return "error/form";
+    }
+
+    @GetMapping("/invalid-delete")
+    public String showInvalidDeletePage(Model model) {
+        model.addAttribute("errorMsg", "댓글이 없는 경우만 삭제가 가능합니다.");
         return "error/form";
     }
 }
