+++ src/main/java/codesquad/springcafe/service/impl/ArticleManagementService.java
@@ -2,7 +2,6 @@ package codesquad.springcafe.service.impl;
 
 import codesquad.springcafe.exception.db.ArticleNotFoundException;
 import codesquad.springcafe.exception.service.DuplicateArticleIdException;
-import codesquad.springcafe.exception.service.DuplicateUserIdException;
 import codesquad.springcafe.model.Article;
 import codesquad.springcafe.model.ListArticle;
 import codesquad.springcafe.model.UpdatedArticle;
@@ -29,14 +28,14 @@ public class ArticleManagementService implements ArticleService {
     }
 
     @Override
-    public void addArticle(Article article) throws DuplicateUserIdException {
+    public void addArticle(Article article) throws DuplicateArticleIdException {
         try {
             validateDuplicateArticleId(article); // 중복 검증
             articleRepository.addArticle(article);
             logger.info("[게시글 생성 완료] - " + article);
         } catch (DuplicateArticleIdException e) {
             logger.error("이미 중복된 ID를 가진 게시글이 존재합니다.");
-            throw new DuplicateUserIdException(article.getUserId());
+            throw new DuplicateArticleIdException(article.getId());
         }
     }
 
@@ -54,17 +53,30 @@ public class ArticleManagementService implements ArticleService {
     @Override
     public Article findArticleById(long id) {
         Optional<Article> optArticle = articleRepository.findArticleById(id);
-        return optArticle.orElseThrow(() -> new ArticleNotFoundException(id));
+        if (optArticle.isPresent()) {
+            Article article = optArticle.get();
+            if (article.isDeleted()) {
+                throw new ArticleNotFoundException(id);
+            }
+            return article;
+        }
+        throw new ArticleNotFoundException(id);
     }
 
     @Override
     public void updateArticle(long id, UpdatedArticle article) {
+        if (articleRepository.isArticleDeleted(id)) {
+            throw new ArticleNotFoundException(id);
+        }
         articleRepository.modifyArticle(id, article);
         logger.info("[{} 게시글 수정 성공]", id);
     }
 
     @Override
     public void deleteArticle(long id) {
+        if (articleRepository.isArticleDeleted(id)) {
+            throw new ArticleNotFoundException(id);
+        }
         long deletedArticleId = articleRepository.deleteArticle(id);
         logger.info("[{} 게시글 삭제 성공]", deletedArticleId);
     }
@@ -76,8 +88,8 @@ public class ArticleManagementService implements ArticleService {
 
     @Override
     public void increaseViewCount(long id) {
-        long increasedViewCount = articleRepository.increaseViewCount(id);
-        logger.info("[{}번째 게시글 조회수 증가] - 현재 조회수 : {}", id, increasedViewCount);
+        articleRepository.increaseViewCount(id);
+        logger.info("[{}번째 게시글 조회수 증가]", id);
     }
 
     @Override
